requirements.txt
# Python dependencies required for development
blinker
click
distlib
filelock
Flask
Flask-Cors
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
MarkupSafe
marshmallow
packaging
pbr
pip
platformdirs
psycopg2
psycopg2-binary
python-dotenv
SQLAlchemy
stevedore
typing_extensions
virtualenv
virtualenv-clone
virtualenvwrapper
Werkzeug
